requirements.txt
future
emoji
lxml
wget
beautifulsoup4
requests
python-telegram-bot==12.8
sqlalchemy==1.3.23
PyDictionary
psycopg2-binary
feedparser
hachoir
pynewtonmath
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp>=2.2.5
Pillow
CurrencyConverter
jikanpy
speedtest-cli
regex
bleach
markdown2
wikipedia
telethon
telegraph
heroku3
nekos.py
alphabet_detector
pyrate-limiter
cachetools
ujson
pretty_errors
TgCrypto
Pyrogram==1.4.16
youtube-dl
youtube_search
asyncio
dateparser
pymongo
dnspython
secureme
apscheduler
emoji-country-flag
countryinfo
html2text
bs4
opencv-python-headless
bing_image_downloader
search_engine_parser
pytz
pykeyboard
aiofiles
motor
telegram
gpytranslate
yt_dlp
opencv-python-headless
youtube_search_python
python-arq
